+++ lib/app/routes/app_routes.dart
@@ -14,7 +14,7 @@ import 'package:sufi_one/app/modules/public/views/register_view.dart';
 import 'package:sufi_one/app/modules/public/views/splash_view.dart';
 import 'package:sufi_one/app/modules/mobcol/views/mobile_collection_view.dart';
 // import 'package:sufi_one/app/modules/survey/views/newtask_confirm_view.dart';
-import 'package:sufi_one/app/modules/survey/bindings/survey_splash_binding.dart';
+import 'package:sufi_one/app/modules/survey/bindings/splash_binding.dart';
 import 'package:sufi_one/app/modules/survey/views/splash_screen_view.dart';
 // import 'package:sufi_one/app/modules/survey/views/upload_checking_view.dart';
 // import 'package:sufi_one/app/modules/survey/views/process_survey_view.dart';
